backend/chatbot/service-chatbot/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
numpy==1.26.4
faiss-cpu==1.8.0.post1
pypdf==6.7.5
